+++ src/main/java/lotto/view/dto/InputMoneyRequest.java
@@ -7,12 +7,12 @@ import java.util.regex.Pattern;
 public class InputMoneyRequest {
     private static final Pattern NUMBER_REGEX = Pattern.compile("^[0-9]*$");
 
-    private final int inputMoney;
+    private final long inputMoney;
 
 
     public InputMoneyRequest(final String inputMoney) {
         validate(inputMoney);
-        this.inputMoney = Integer.parseInt(inputMoney);
+        this.inputMoney = Long.parseLong(inputMoney);
     }
 
     private void validate(final String lottoMoney) {
@@ -29,7 +29,7 @@ public class InputMoneyRequest {
         return NUMBER_REGEX.matcher(lottoMoney).matches();
     }
 
-    public int calculateLottoCount(final int lottoPrice) {
+    public long calculateLottoCount(final long lottoPrice) {
         return inputMoney / lottoPrice;
     }
